Rename to avoid name clash on rubygems.org
serverspec_extensions -> tapi-serverspec
