--- lib/serverspec_extensions.rb
@@ -1,2 +1,0 @@
-module ServerspecExtensions
-end
