--- lib/serverspec_extensions/matcher.rb
@@ -1,1 +1,0 @@
-require 'serverspec_extensions/matcher/be_installed_as_java'
