--- lib/serverspec_extensions/type.rb
@@ -1,6 +1,0 @@
-require 'serverspec_extensions/type/java'
-
-module ServerspecExtensions
-  module Type
-  end
-end
