--- lib/serverspec_extensions/version.rb
@@ -1,3 +1,0 @@
-module ServerspecExtensions
-  VERSION = '0.1.0'
-end
