+++ lib/tapi/serverspec.rb
@@ -1,0 +1,4 @@
+module TAPI
+  module Serverspec
+  end
+end
