+++ lib/tapi/serverspec/matcher.rb
@@ -1,0 +1,1 @@
+require 'tapi/serverspec/matcher/be_installed_as_java'
