+++ lib/tapi/serverspec/type.rb
@@ -1,0 +1,8 @@
+require 'tapi/serverspec/type/java'
+
+module TAPI
+  module Serverspec
+    module Type
+    end
+  end
+end
