+++ lib/tapi/serverspec/version.rb
@@ -1,0 +1,5 @@
+module TAPI
+  module Serverspec
+    VERSION = '0.1.0'
+  end
+end
